Update patch set 1
Patch Set 1:

REQUIRE is needed for the I64Mul opcode when run on a 32-bit MIPS architecture, since in that context the implementation is (currently) incorrect.

My read on the compilation failures was that I forgot to add a REQUIRE for uses of I64UConvertF64.

What I wasn't sure of is when and where to add REQUIRE, and hence what elements should go into FOREACH_I64_OPERATOR. I see two choices:

1) We should have all I64 operation in this macro, and REQUIRE when used (making it easier to deal with an architecture that can't implement an operator correctly).

2) Only operations that are broken should be put into the FOREACH_I64_OPERATOR, and all test uses should be fixed to use REQUIRE.

I initially thought case 1) was the better choice, but that assumes that people actually add REQUIRE macros in all use cases. This assumption is bad in that there is nothing to show that one has left out a REQUIRE. Hence, my current thinking is to only use REQUIRE on operations that are broken, and require the person who wants to turn off an operation puts the necessary REQUIRE's in the code.

This would reduce the list of operations needing REQUIRE to I64Mul.

Does this solution sound like the best approach, or do you have a different suggestion?

Patch-set: 1
Reviewer: Gerrit User 1224119 <1224119@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0