[builtins] Turn NewPromiseCapability into a proper builtin.
Instead of inlining NewPromiseCapability everywhere, we should have it
as a separate builtin that can be called. The motivation for doing so is
to reduce the overhead of the builtins and just share the slow-path
handling in a single place, plus this way we can also utilize the
NewPromiseCapability functionality from TurboFan optimized code
eventually.

Also remove the unused %new_promise_capability intrinsic and the builtin
backing it.

R=yangguo@chromium.org

Bug: v8:7253
Change-Id: Ic7c916a11f12d063a645e2c25cc4120badee5640