Update patch set 4
Patch Set 4:

Changed the use of the REQUIRE() macro to only be necessary if the operation (i.e. opcode) is
not supported on some architecture. However, unlike the previous definition, it can be used even if it is supported on all architectures.

This CL is ready for another review.

Patch-set: 4
Reviewer: Gerrit User 1224119 <1224119@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>