Update patch set 1
Patch Set 1: Commit-Queue+2

(1 comment)

This is a general refactoring I wanted to do, and it also fixes an issue with the next CL to allow bailout from assemblers. The problem was that parameters are sometimes passed in registers which are not cache registers, but we stored them in a LiftoffRegister anyway. This lead to DCHECK errors. This mostly happens because we now go through compilation also for unsupported platforms, but it was wrong all the time, so I just fixed it :)

Patch-set: 1
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2