Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > Normally such changes should go through a deprecation period, but since this is wasm I'm guessing it's not widely-used. But in case there's some embedder using this API, can you add to the CL description what the replacement for this is?
> 
> Thanks for the reminder, I forget sometimes that there are other embedders beside Chrome and node.js. I adjusted the CL description now.
> 
> How long is the deprecation period?

The historical deprecation period is one release with V8_DEPRECATE_SOON, then a release with V8_DEPRECATE, then deleted. But that was designed around things with much less...functionality than this. I can understand that it might be exceedingly painful to support a dead API for three releases (18 weeks) at the least.

Patch-set: 3