Update patch set 2
Patch Set 2:

I would still remove it. The fact that it accidentally helps doxbee is not a good basis for such optimizations, especially when they are inconsistent between Ignition and TurboFan.

Let's figure out how this affects doxbee exactly and then see if we can do something about it. Otherwise let's just swallow the regression for now.

Patch-set: 2