Update patch set 5
Patch Set 5:

After experimenting with local chrome build I see that we need to have "Detached" in the name of a node. This way the user can search for "Detached" and find all the detached nodes.

If we have this in the name, then there is no need to a new HeapEntry::kDetached type.

Patch-set: 5
Reviewer: Gerrit User 1002405 <1002405@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>