Relieve worker threads when pausing concurrent marking.
This will allow using more workers for concurrent marking instead of
having to keep a few workers behind for a potential future main thread
task.
Task count will be increased in a follow-up: https://chromium-review.googlesource.com/c/v8/v8/+/918663

Note: this makes it such that previous callers of ConcurrentMarking::EnsureCompleted()
now wait for all tasks instead of trying to cancel unstarted tasks. I think this
makes sense as if you're going to wait on one task you might as well keep
up the parallelism to finish ASAP.

The two pieces of code affected by this are:

    if (FLAG_parallel_marking) {
      DCHECK(FLAG_concurrent_marking);
      heap_->concurrent_marking()->RescheduleTasksIfNeeded();
    }
    ProcessMarkingWorklist();

    FinishConcurrentMarking();

(clearly this didn't intend for FinishConcurrentMarking() to immediately cancel tasks it
just started)

and

  if (was_marked_incrementally_ && heap_->ShouldAbortIncrementalMarking()) {
    heap()->incremental_marking()->Stop();
    heap()->incremental_marking()->AbortBlackAllocation();
    FinishConcurrentMarking();
    heap()->incremental_marking()->Deactivate();
    (...)

I will follow-up with another change to make this use PREEMPT_TASKS (not doing it
in this CL since it may have unknown side-effects).


Bug: chromium:812178
Change-Id: I16ef90295fadaac82e031f204b0a7a8cd0931524