Update patch set 5
Patch Set 5:

On further thought, I'm not sure if this is ready to review.

My process was to look at a list of builtins [1] and add whitelist ones that would pass the tests.  I realize that I'm not familiar with what operations are unsafe, e.g. 'CreateDataProperty', so I'm looking for guidance on how to determine which ones are actually side-effect-free

[1] https://cs.chromium.org/chromium/src/v8/src/builtins/builtins-definitions.h

Patch-set: 5
Reviewer: Gerrit User 1230559 <1230559@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>