[turbofan] Reset the speculation poison register on IfException.
While implementing load poisoning on Arm64, we were getting spurious crashes
when running web-tooling. It turned to be caused by the poison not being reset
when a call would return to a block starting with "IfException".

This patch introduces a new bit to mark instructions as "Catch", in the same way
that we mark instructions as "Call". We can then use this bit to reset the
poison register.

Change-Id: I3cd520b73415fac1bbcd2446ecf662bb821666f3