Create a new read-only space
Adds a new space RO_SPACE and modifies the serializer and other machinery to
support it.

Currently the new space has nothing in it, but will eventually have all the
immortal immovable immutable objects in it, so the GC can ignore the space and
later still the space will not even be serialized but mapped directly into
memory.

Cq-Include-Trybots: master.tryserver.chromium.linux:linux_chromium_rel_ng
Change-Id: Ib2ff474699196c138df8c24f7a2248471e30fbac