[turbofan] Collect and use SharedFunctionInfo for inlining/calling.
Change the CALL_IC logic in Ignition to collect the SharedFunctionInfo
(actually the feedback vector cell from the closures, which gives us
access to both SharedFunctionInfo and FeedbackVector, and requires only
a single comparison to also guard against cross-context inlining) for
polymorphic call sites.

Introduce a new CheckClosure operator into TurboFan to handle the
dynamic checks in optimized code. This operator carries the feedback
vector slot from the closure.

On the micro-benchmark initially proposed by Slava (on the tracking bug)
we go from

  console.timeEnd: first, 61.098000
  console.timeEnd: second, 995.684000
  console.timeEnd: third, 527.001000

to

  console.timeEnd: first, 61.309000
  console.timeEnd: second, 212.798000
  console.timeEnd: third, 106.130000

which corresponds roughly to a ~5x performance boost. On the SixSpeed
test suite the generator-es5 benchmark improves from 143ms to 58ms,
which corresponds to a 2.5x improvement.

Bug: v8:2206, v8:7253
Change-Id: I37227a8a83765eca972dd9680d8551de6af88ed9