Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm also working on an arm64 port. It's not uploaded yet, but I hope to have it fully working tomorrow. I've discovered some interesting things, for example tbz/cbz instructions have to be avoided for branches with our system based on flags. If you like, I'll let you know as it's finished, and maybe this will be useful for your research too?

That would be great yes! I'll be happy to review it. I came to the same conclusion regarding tbz and cbz, that's a little sad but I'm afraid we might not have a choice. The alternative being:

  cbz x0, <B>
  cmp x0, #0
  csel x18, xzr, x18, eq

B:
  cmp x0, #0
  csel x18, xzr, x18, ne

Which defeats the idea of the cbz all together.

Patch-set: 1
Reviewer: Gerrit User 1183769 <1183769@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0