[cleanup] Avoid redundant `TaggedIsSmi` checks in CSA
`CodeStubAssembler::ToInteger_Inline` performs a `TaggedIsSmi` check
before calling `ToInteger` with the appropriate truncation mode.

When we already know we’re dealing with something that’s either a Smi
or not a Smi, this check is redundant. This patch bypasses the check
by using `CallBuiltin(Builtins::kToInteger*)` directly in such cases.

Bug: v8:7310
Change-Id: If538e39bcb738014bd03f10edd0051dac72b7ea3