[builtins] Stack unwinding support for off-heap builtins
When unwinding the stack, we use the relative pc position to index
into the handler table. This adds support for off-heap code, for which
the on-heap code object only contains the trampoline.

In a follow-up, we should add new accessors to Code to encapsulate
this logic.

Drive-by-fix: DCHECK in LookupCode.

Bug: v8:6666
Change-Id: I30bee9087d9d654e36ed968fc43a76b6ed90ddc7