[esnext] re-implement template strings
- Add a new bytecode for the ToString operation, replacing the old
intrinsic call (TODO: consider using feedback slots for the load
and call to speed up benchmarks --- but this is probably not useful
in real world applications)
- Add a new AST node to represent TemplateLiterals, and avoid
generating unnecessary ToString operations (in some simple cases),
and avoid unnecessary string concatenation in effect contexts.
- Use a single feedback slot for each string addition, because the
type feedback should always be the same for each addition

This seems to produce a very slight improvement on JSTests benchmarks
and bench-ruben.js from v8:7415, and it's possible that type feedback
for the ToString bytecode could provide more opportunities to eliminate
the runtime call in TurboFan.

Doesn't touch tagged templates

BUG=v8:7415
R=rmcilroy@chromium.org, ishell@chromium.org, bmeurer@chromium.org

Change-Id: If5a8c68558431f058db894d65776324abf54218e