[bigint] Require --harmony-bigint for deserialization
While deserializing a BigInt with the --harmony-bigint flag off is
harmless in itself, trying to wrap one as an Object (either during
deserialization of a JSValue, or later from user code) will crash.
To avoid that risk, this patch makes  the deserializer will reject
data containing a BigInt or BigIntObject when the flag is off.

Bug: chromium:820819
Change-Id: I024a4f13715bbe95ee8eb6e1710e8f47ca227644