Update patch set 16
Patch Set 16:

> Patch Set 16: Commit-Queue+1
> 
> I reduced exposure of CheckSideEffect in generator to one line per instrumented bytecode. I prefer this solution because it is more explicit then PreHook, provide minimum performance effect based on measurement from doc and takes only one line for instrumentation.
> Yang and Ross, WDYT?

I think this is precisely what Ross did not want. This way, you would have to add CheckSideEffect in many places, and it becomes hard to get an overview. Having it in the macro and add the prehook based on a list sounds a lot more appealing. That list can then also be used in debug-evaluate.cc to ensure that BytecodeHasNoSideEffect only whitelists store bytecodes that have the prehook

Patch-set: 16
Reviewer: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>