Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Curious. Im what case would two siblings have the same start and end?

It happens when a singleton range is extended. The originally produced slots are:

 {112,203}  // 1 then
 {203,303}  // 2 else
 {166,-1}   // 3 return 'if'; continuation
 {203,-1}   // 4 then continuation
 {268,-1}   // 5 return 'else'; continuation
 {303,-1}   // 6 else continuation 
 {303,-1}   // 7 ... i'd have to double-check what this one is.

What happens is that the the singletons starting at 203 and 268 (both have a count of 0) are merged. This results in a {203, 303} range with count = 0.

It's a bit unfortunate that the then continuation does not terminate with the beginning of the else block. This used to be the case at some point, until https://chromium-review.googlesource.com/816736 when we extended the 'else' range to also include the 'else' keyword.

Patch-set: 3
Reviewer: Gerrit User 1162439 <1162439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>