Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> What is the plan of moving forward with this? It is fairly unusual in V8 to land an API change without implementation and tests.

My thoughts on this:

The rough plan was to get feedback on this and either land it as-is, or combine it with a rewritten WASM module serialization using this API. Combining this with this CL souldn't be that much larger (just two more files, wasm-serialization.*) and perhaps some test files. I'm not sure how many would need to be updated.

WASM module serialization is different from normal serialization, in that the buffers may have to be very large (512 MB or more), and the embedder may want to do different things with the pieces (caching code and the portable wire bytes separately, etc.) so adding methods to the serialization delegates seems like a reasonable choice. An alternative design would be to define new standalone APIs, say ModuleSerializer/Deserializer. The key requirements are for V8 to support large modules and to separate the different module pieces.

Patch-set: 3
Reviewer: Gerrit User 1001213 <1001213@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>