Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Is this a quick fix for the crash mentioned in the bug? Can you update the CL description to match that?

Ok. Will do. 
> 
> Are there any blockers to implementing the spec?

Actually, we need a bit of discussion as to the merit of the current spec as mentioned in the bug comment (Jack from MS also thinks along this line). It seems a bit odd to expect the calendar to be 'Chinese' (year, month, day are from Chinese calendar) while picking up a pattern for formatting from the default calendar of the locale (in most case, Gregorian).

Patch-set: 5
Reviewer: Gerrit User 1151758 <1151758@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>